requirements.txt
pillow
numpy
docopt
scipy
scikit-image